pyproject.toml
[project]
name = "django-headless"
version = "0.1.0"
description = "Automagically create an API for you Django models"
authors = [
    {name = "Leon van der Grient",email = "leon@devtastic.io"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "rich (>=14.0.0)",
]
optional-dependencies = [
    "drf-flex-fields (>=1.0.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.development.dependencies]
black = "^25.1.0"
djangorestframework = "^3.16.0"
django = "^5.2.4"

